feat(xudctl): add launch checks
Based on xud-simnet's `xud-simnet-channels` script. It will check whether BTC and LTC channels is open or not. If channels are not open yet then xudctl will block and wait until channels are ready. In the meantime xudctl will print dots in screen like xud-simnet. Users can type `Ctrl-C` to stop waiting but next time when users execute xud script and if the channels are still not ready then dots will continute printing.