fix: gracefully shut down containers
In this commit we fix an issue where geth, litcoind, xud, ltcd-simnet and raiden were
not shutting down gracefully.

The global timeout for graceful shutdown is 180 seconds which should be
enough for mainnet on devices with low resources. Once the timeout is
reached the containers are killed forcefully.

I also added a short 10 second delay before starting up raiden so that
geth has time to boot up. It will serve as a quick, temporary fix until
we add a more reliable solution.