feat: support external full nodes
This commit adds configuration files (toml format) to xud-docker. A new container named "exchangeunion/utils" is added to replace most of the bash scripts outside of the docker. So docker and bash will be the only two dependencies after this commit. Configuration files are created automatically when you are going to launch any networks. You can use these files to customize bitcoind, litecoind and geth at the moment.