feat(xudctl): add confirmation before down command (#483)
* feat(xudctl): add confirmation before down command

* exit xudctl after down but drop out confirmation

It's becuase it's not an easy task to add interactive question to
commands running inside xudctl. The deeper reason is that the xudctl
shell is running a sync-mode. That means key handling and command
executing cannot happen in the same time. That's the same reason behind
problematic Ctrl+C a long time ago.